services/sensor_reader/requirements.txt
paho-mqtt
PyYAML
adafruit-circuitpython-ahtx0
adafruit-circuitpython-bmp280